more work. current issues: ninian glacies doesn't calc damage properly, reworked ardent aspect doesn't seem to work. relic images are non centered?

+++ theDefault/src/main/java/SNAct1/powers/SoulChargePower.java
@@ -64,6 +64,6 @@ public class SoulChargePower extends TwoAmountPower {
         if (type != DamageInfo.DamageType.NORMAL) {
             return damage;
         }
-        return damage * (1 + (amount2 / 100));
+        return damage * (1 + ((float)amount2 / 100f));
     }
 }
